Moving more functions into mongo package
